login signup conn to api

+++ yoke-server/src/auth/auth.controller.ts
@@ -9,10 +9,11 @@ import {
   HttpException,
   HttpStatus,
   HttpCode,
+  Req,
 } from '@nestjs/common';
 import HashEncrypt from 'src/common/utils/HashEncrypt';
 import { AuthSchema, option } from 'src/common/utils/Joi';
-import { IUser, ICreateUser, IUpdateUser } from '../common/interfaces/user';
+import { IUser, ICreateUser } from '../common/interfaces/user';
 import { AuthService } from './auth.service';
 
 @Controller('auth')
@@ -28,11 +29,15 @@ export class AuthController {
   async findOne(
     @Param('id') id: string,
   ): Promise<{ user: IUser; message: string }> {
-    const user = (await this.authService.findOne(id)) as unknown as IUser;
-    if (!user) {
-      throw new HttpException('User not found.', HttpStatus.NOT_FOUND);
+    try {
+      const user = (await this.authService.findOne(id)) as unknown as IUser;
+      if (!user) {
+        throw new HttpException('User not found.', HttpStatus.NOT_FOUND);
+      }
+      return { user, message: 'successful' };
+    } catch (err) {
+      console.log(err);
     }
-    return { user, message: 'successful' };
   }
 
   @Post('/signup')
@@ -48,8 +53,6 @@ export class AuthController {
       );
     }
 
-    const hashedPassword = await HashEncrypt.hashPassword(user.password);
-    user.password = hashedPassword;
     const existingUser = (await this.authService.findByEmail(
       user.email,
     )) as unknown as IUser;
@@ -59,6 +62,8 @@ export class AuthController {
         HttpStatus.NOT_ACCEPTABLE,
       );
     }
+    const hashedPassword = await HashEncrypt.hashPassword(user.password);
+    user.password = hashedPassword;
 
     const newUser = (await this.authService.createOne(
       user,
@@ -66,9 +71,18 @@ export class AuthController {
     return { user: newUser, message: 'Signup successful' };
   }
 
-  @Delete()
+  @Delete(':id')
   @HttpCode(204)
-  async deleteUser(@Param() id: string): Promise<{ message: string }> {
+  async deleteUser(
+    @Req() req,
+    @Param('id') id: string,
+  ): Promise<{ message: string }> {
+    if (id !== req.user.id) {
+      throw new HttpException(
+        'You can only delete your account',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
     await this.authService.deleteOne(id);
     return { message: 'delete successfully' };
   }
@@ -84,7 +98,7 @@ export class AuthController {
 
   @Post('login')
   async loginUser(
-    user: IUser,
+    @Body() user: IUser,
   ): Promise<{ user: IUser; token: string; message: string }> {
     const validateResult = AuthSchema.validate(user, option);
 
@@ -97,14 +111,20 @@ export class AuthController {
 
     const existingUser = await this.authService.findByEmail(user.email);
     if (!existingUser) {
-      throw new HttpException('Invalid email', HttpStatus.NOT_FOUND);
+      throw new HttpException(
+        'Invalid email or password',
+        HttpStatus.NOT_FOUND,
+      );
     }
     const isAuthenticated = await HashEncrypt.verifyPassword(
       user.password,
       existingUser.password,
     );
     if (!isAuthenticated) {
-      throw new HttpException('Invalid password', HttpStatus.UNAUTHORIZED);
+      throw new HttpException(
+        'Invalid email or password',
+        HttpStatus.NOT_FOUND,
+      );
     } else {
       const token = HashEncrypt.generateToken({
         id: existingUser.id,
